requirements.txt
github-copilot-sdk>=0.1.25
pydantic>=2.0
rich>=13.0
python-dotenv>=1.0
pyyaml>=6.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
msal>=1.31.0
itsdangerous>=2.2.0
azure-identity>=1.19.0
azure-mgmt-resource>=23.0.0
pyodbc>=5.1.0
requests>=2.31.0
